% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microarray.R
\name{append_multiprobe_genes}
\alias{append_multiprobe_genes}
\title{Summarised gene symbols by probeset}
\usage{
append_multiprobe_genes(df, gene = gene_symbol, probe = probeset_id)
}
\arguments{
\item{df}{data.frame or tibble}

\item{gene}{unquotted column name for gene symbols}

\item{probe}{unquotted column name for probeset_id}
}
\value{
A data.frame of summarised gene symbols
}
\description{
Append probeset ids when more than 1 for gene symbol
the returned data frame will have unique gene symbols present only once
}
\examples{
\dontrun{
library(bifag)

res <- append_multiprobe_genes(df)
}

}
