% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{cluster_mount}
\alias{cluster_mount}
\title{Connect to a cluster using sshfs
test if the cluster is already mounted}
\usage{
cluster_mount(cluster)
}
\arguments{
\item{cluster}{cluster name}
}
\description{
Connect to a cluster using sshfs
test if the cluster is already mounted
}
