% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{dates_report}
\alias{dates_report}
\title{fill dates for Rmd header, with creation and last modification date}
\usage{
dates_report(input = NULL)
}
\arguments{
\item{input}{vector of an input file, if null use current Rmd from knitting}
}
\value{
character string of created/modif/today dates for Rmd header date field
}
\description{
fill dates for Rmd header, with creation and last modification date
}
