% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rna-seq.R
\name{n_targets}
\alias{n_targets}
\title{Output number of retained targets according to p-value and logFC threshold}
\usage{
n_targets(.data, pval = 0.05, logratio = 1)
}
\arguments{
\item{.data}{data.frame or tibble}

\item{pval}{adjusted p-value threshold}

\item{logratio}{log ratio threhold}
}
\value{
number of retained targets
}
\description{
for DESeq2 results input tables
}
