% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{str_subset_inv}
\alias{str_subset_inv}
\title{inverse version of str_subset}
\usage{
str_subset_inv(vec, pattern)
}
\arguments{
\item{vec}{a vector}

\item{pattern}{pattern as character}
}
\value{
a vector without elements containing the pattern
}
\description{
inverse version of str_subset
}
