% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{cluster_mount}
\alias{cluster_mount}
\title{test if any cluster is already mounted and do it if not}
\usage{
cluster_mount(cluster)
}
\arguments{
\item{cluster}{cluster name}
}
\description{
test if any cluster is already mounted and do it if not
}
