% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{session_info_nodep}
\alias{session_info_nodep}
\title{look from library calls and display their version and origin in a tibble
From Eric Koncina}
\usage{
session_info_nodep(file = NULL, type = c("all", "library"))
}
\arguments{
\item{file}{NULL for current doc or a path to a file}

\item{type}{either 'library' or 'colons' for library calls or colons for '::' calls}
}
\value{
A tibble of used packages
}
\description{
look from library calls and display their version and origin in a tibble
From Eric Koncina
}
